Changes added:
add-unix.pl -
Updated to allow reference to a passfile  (-p option). The passfile argument is expected to be a full path name to a file (or named pipe, etc) which contains the *unencrypted* user password to be used. This method of storing a password for an account is insecure and should be replaced with a secure keychain access method. Note that the -p option can only be used along with the -a (altuser) option at this time.

Also fixed a bug in the hostname/domain name splitting logic which failed to properly separate the short hostname from a (sub)domain name with more than 2 dots in it. (I.e, myhost.domain.com was fine, but myhost.project.region.domain.com would not parse well.)

Audit.pm -
Made output from workerThread a little less noisy when DEBUG mode isn’t on. Fixed lookupHostDNS’s hostname/domain name splitting logic similarly to add-unix.pl to handle longer subdomain chains.

Added support for using sshpass plus a stored user password to login to machines which do not have preinstalled SSH keys. Note that this method of storing passwords in a local (unencrypted) file is not considered a secure practice, and thus it’s wise to use this method only as a bootstrapping process until you have installed working SSH keys onto your nodes, and then you can remove this stuff.

GauntletWeb.pm -
Added support for generating XLS format report files in ViewResults(). In the process, removed the link to download a CSV file. The reasoning for this is because I found problems with the CSV generation, specifically in its method of encoding output which itself includes commas and other characters which may break the current CSV encoding method.

TaskMan.pm -
Added support for passfile feature to getHostOverrides().